refactor: factor out version determine, build YAML, build Markdown and data mutation checks
